Compiling framework!
It's a little sparse, but the organization is all there.

Pattern is an abstract class, which is implemented in RgbVerticalPattern
(and others in the future).  Each pattern has a number of widgets
associated with it, which are also abstract and implemented with
ThreeWheelWidget (and others in the future).  Each derived Widget
has a number of channels in it, implemented in WidgetChannel (this will
be expanded further).