Add cscope.out to files git should ignore.
cscope.out is a reference file for cscope, which allows navigation
to and from function declarations from where they are called in other
files using Ctrl + ] and Ctrl + t.

To generate cscope output, use:

cscope -Rb

Then, when in Vi, place the cursor over a function you would like
to see the declaration for and press Ctrl + ].  Return from it
using Ctrl + t.