Updates throughout to start bringing in a priority scheme.
A "frame" contains the final OPC data to be sent to the opc-server.
It is built by passing each pattern that has been updated since the
last loop along with its priority into buildFrame(), which determines
how to update the pixel values based on the provided pattern's
values and priority.

For now it's a direct replacement - each priority overwrites the
entire array with its Pattern's values.

Did some other minor stuff:
    - Added a "name" string to the Pattern superclass
    - Assigned the name in the Pattern's derived class initPattern
      function
    - Added a "numChannels" integer to the Widget superclass
    - Assigned the numChannels in the Widget's derived class initWidget
      function
    - Added a printInit() function to print the initialized state of a
      pattern after initPattern() and initWidgets()