Adding the version of GardenSpinner from ~/devl/GardenSpinner_temp on perry.
I don't remember why it exists, but the git log shows some changes, and there
are changes that aren't in the "official" version.  The changes were done at
almost the end of The Beforetime.  I don't know if they were experimental or
if they were what we really ran in the Rose Garden at WGaG in 2019.